start validator
