+++ src/components/constants.js
@@ -1,0 +1,11 @@
+//Валидация форм
+const settings = {
+    formSelector: ".popup__edit",
+    inputSelector: ".popup__info",
+    submitButtonSelector: ".popup__save",
+    inactiveButtonClass: "popup__save_inactive",
+    inputErrorClass: "popup__edit_type_error",
+    errorClass: "popup__error-massage_active",
+  };
+
+export {settings}
